feat: refactor Menu Items and Menus
