--- src/menu/login.ts
@@ -1,13 +1,0 @@
-import type { Menu } from 'types/provideTypes';
-import * as Item from './menuItems';
-
-export const headerOut: Menu = [
-    Item.mainPage,
-    Item.register,
-];
-
-export const footerOut: Menu = [
-    Item.register,
-    Item.contact,
-    Item.imprint,
-];
