--- src/menu/verify.ts
@@ -1,14 +1,0 @@
-import type { Menu } from "types/provideTypes";
-import * as Item from './menuItems'
-
-export const header: Menu = [
-    Item.mainPage,
-    Item.login,
-    Item.contact,
-];
-
-export const footer: Menu = [
-    Item.mainPage,
-    Item.login,
-    Item.contact,
-];
