+++ src/menu/yearItems.ts
@@ -1,0 +1,25 @@
+import type { MenuItem } from 'types/provideTypes';
+
+export const speaker: MenuItem = {
+    name:        'Vortragende',
+    url:         '#Speaker',
+    description: 'Klicke, um zu den Vortragenden zu navigieren',
+};
+
+export const sponsors: MenuItem = {
+    name:        'Partner',
+    url:         '#Sponsors',
+    description: 'Klicke, um zu den Sponsoren und Partnern zu navigieren',
+};
+
+export const team: MenuItem = {
+    name:        'Team',
+    url:         '#Team',
+    description: 'Klicke, um zum Tech Stream Conference-Team zu navigieren',
+};
+
+export const schedule: MenuItem = {
+    name:        'Ablaufplan',
+    url:         '#Schedule',
+    description: 'Klicke, um zum Ablaufplan der Konferenz zu navigieren',
+};
